+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 2.x branch
 ## 2.6 branch
 ### 2.6.1
+* added `DriverTestCase::getMockForDriver()` method;
+* `DriverTestCase::allRecords()` method renamed as `getAllRecords()`;
 * updated for `php-tools` 1.1.7.
 
 ### 2.6.0